frontend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
mysql-connector-python==8.2.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.4.0
